# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iTaxCix API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
        "/admin/permission/create": {
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Crear permiso",
                "description": "Crea un nuevo permiso en el sistema.",
                "operationId": "createPermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PermissionCreateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permiso creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Creado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Petición inválida"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Campos requeridos faltantes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token inválido o expirado"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/permission/update": {
            "put": {
                "tags": [
                    "Permission"
                ],
                "summary": "Actualizar permiso",
                "description": "Actualiza un permiso existente.",
                "operationId": "updatePermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PermissionUpdateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/permission/delete": {
            "delete": {
                "tags": [
                    "Permission"
                ],
                "summary": "Eliminar permiso",
                "description": "Elimina un permiso del sistema.",
                "operationId": "deletePermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PermissionDeleteRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Permiso eliminado exitosamente (sin contenido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/permission/list": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Listar permisos",
                "description": "Lista todos los permisos del sistema.",
                "operationId": "listPermissions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/PermissionResponseDTO"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "perPage": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "currentPage": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "lastPage": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PermissionCreateRequestDTO": {
                "description": "Datos para crear un permiso",
                "properties": {
                    "name": {
                        "description": "Nombre del permiso",
                        "type": "string",
                        "example": "Ver usuarios"
                    },
                    "active": {
                        "description": "Si el permiso está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "web": {
                        "description": "Si el permiso es para web",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "PermissionDeleteRequestDTO": {
                "description": "Datos para eliminar un permiso",
                "properties": {
                    "id": {
                        "description": "ID del permiso",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "PermissionResponseDTO": {
                "description": "Datos de respuesta de permiso",
                "properties": {
                    "id": {
                        "description": "ID del permiso",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del permiso",
                        "type": "string",
                        "example": "Ver usuarios"
                    },
                    "active": {
                        "description": "Si el permiso está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "web": {
                        "description": "Si el permiso es para web",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "PermissionUpdateRequestDTO": {
                "description": "Datos para actualizar un permiso",
                "properties": {
                    "id": {
                        "description": "ID del permiso",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del permiso",
                        "type": "string",
                        "example": "Ver usuarios"
                    },
                    "active": {
                        "description": "Si el permiso está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "web": {
                        "description": "Si el permiso es para web",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}